Completed Disassembler.
Instruction information is saved instead of strings. Upcoming instructions are loaded into the future list.

Removed run_for_cycles() from the emulator Core. step_instruction() is now used instead.

Added check to only fill the disassembler history FIFO and future list if debug tools were enabled.

+++ Src/Cores/GB/SM83.hpp
@@ -74,7 +74,7 @@ namespace GB
         uint16_t sp = 0xFFFE, pc = 0;
         std::array<uint8_t, 8> registers{};
 
-        size_t fetched_count = 0;
+        uint8_t fetched_count = 0;
         std::array<uint8_t, 3> fetched_bytes{};
 
         SM83(Core &core, MainBus &bus);
